Audiofiles Update
Update Audiofile to send data to device
send next data file orr previous according the message received from device.
Able to end each file.
Update progress of each file
Update ovarall progress
add two data sizes in wificonfig

+++ Booyco HMI Utility/Bootloader/Configure/AudioFiles/AudioEntry.cs
@@ -79,8 +79,8 @@ namespace Booyco_HMI_Utility
             }
         }
 
-        private UInt16 _Progress;
-        public UInt16 Progress
+        private int _Progress;
+        public int Progress
         {
             get
             {
